vote-interface: move V4 tests into vote_state_v4.rs (#591)
